Support for --web when using gist create (#2263)
* Support for --web when using gist create

Proposal to close #2071

I have not worked with Go prior to this so please smite me down with the
wisdom of a million Golang gods if I'm doing something terribly wrong.

I also added a test to gist/create for the added web arg.

Pretty much referenced the implementation from pr/create.

* Fix for Tests / build (windows-latest)

I believe this fixes it as it stopped failing on a local vm. Otherwise I
will try and tackle it tomorrow.

* minor cleanup

Co-authored-by: vilmibm <vilmibm@github.com>